add loading state to nft detail page
